pre-process model added to API

+++ brahmi_backend/controller.py
@@ -4,6 +4,7 @@ import filetype
 import services
 from util import make_response
 from segmentation import image_segmentation
+from pre_process_module import preprocessImage
 
 app = Flask(__name__)
 
@@ -21,6 +22,25 @@ def translateLetters():
             classify_letters = services.classify_letters()
             result = {'letter': classify_letters}
             response = make_response(result,True,200)
+            os.remove(os.path.join(input_data, image_name))
+            return Response(response=response, status=200, mimetype='application/json')
+        else:
+            response = make_response('The file is NOT an Image', False, 200)
+            return Response(response=response, status=200, mimetype='application/json')
+    except Exception as e:
+        print(e)
+        response = make_response('The file is NOT FOUND', False, 404)
+        return Response(response=response, status=404, mimetype='application/json')
+
+@app.route("/api/preprocessImage", methods=["POST"])
+def prePrecessImage():
+    try:
+        image = request.files["image"]
+        image_name = image.filename
+        image.save(os.path.join(input_data, image_name))
+        if filetype.is_image(os.path.join(input_data, image_name)):
+            preprocessImage(image_name)
+            response = make_response("Pre Processing Complete", True, 200)
             os.remove(os.path.join(input_data, image_name))
             return Response(response=response, status=200, mimetype='application/json')
         else:
